i#1736 better options: add new dr_client_main() init routine
Deprecates dr_init(), replacing it with dr_client_main(), which takes in
argc and argv just like a standalone main() routine.
DR supports both, for backward compatibility.

Updates all references in the docs and in our samples.  Updates one test
but leaves the rest of the tests and our end-user tools for later.

Review-URL: https://codereview.appspot.com/253160044