i#1714 windows 10: update winsysnums for new WOW64 gateway
Updates winsysnums to decode Windows 10 WOW64 gateways, using a check
for the computed preferred base to ensure the immediate is within the
library (imagehlp doesn't load the library at its preferred base and
doesn't provide the preferred base).

Enables winsysnums to be built by default in a regular DR build.

Review-URL: https://codereview.appspot.com/260760043