i#511 drreg: support reserved regs across app instrs
Adds handling of app reads and writes while registers are reserved.  On a
read, restores the app value for the app instr and saves the tool value in
a separate TLS slot; on a write, does the same but re-spills the app
value.  The recommend TLS slot count is documented to include an extra slot
for each simultaneous register reserved across app instrs.

Adds testing via labeled basic block sequences using duplicate immediates
for targeted tests.

Review-URL: https://codereview.appspot.com/269810043