i#511 drreg: add fault handling
Adds register restoring on a fault.  The approach is to walk the cache code
and identify spills and restores, including distinguishing tool spills from
app spills.

Adds tests to drreg-test, including that a spilled register is restored to
the app value for the app's fault handler.

Ports earlier asm drreg test code to ARM.

Adds checks for ops.conservative to force spilling even when registers are
dead.

Review-URL: https://codereview.appspot.com/261420043