i#58 MacOS: enable initial test suite
Now that the tests build on MacOS, we enable an initial test suite.  We use
a test label "OSX" to select the one-quarter of the tests that currently
run.  i#1815 covers fixing the rest of the tests.

We also add a 32_only arg to runsuite_common and set it for Mac.
Additionally, we increase the default test timeout from 60s to 90s to
support older hardware.

Review-URL: https://codereview.appspot.com/274750043