i#1711: augment drx_insert_counter_update() to work with drreg
Adds support to drx_insert_counter_update() to work with drreg and drmgr.
If called from drmgr's insertion phase, drreg is used to spill the flags on
x86 or to obtain two scratch registers on ARM.

Adds support for drmgr_current_bb_phase() to be called without calling
drmgr_init() as a test for whether drmgr is in use.

Adds two new tests: one of drx_insert_counter_update() without drmgr or
drreg in drx-test, and another with drmgr and drreg in a new client test
drxmgr-test.

Fixes #1711

Review-URL: https://codereview.appspot.com/274140043