i#1818 review auth: use oauth2 authentication and add caching support
Adds oauth2 authentication as the default for uploading review diffs.
Adds tee-ing of the upload.py invocation so the user can see output from
the script while it waits for a browser window launch.

Adds a new option --oauth2_token to upload.py for self-caching of the
token to avoid launching a browser window.
Adds "-k <token>" support to git_review.sh.

Fixes #1818

Review-URL: https://codereview.appspot.com/274950043