i#1845: fix ARM try-fault sigmask assert
Adds a call to dr_setjmp_sigmask() from dr_setjmp(), which was omitted from
the original ARM implementation in 5688245.  This fixes an assert.

Adds a test of DR_TRY to the client.exception test, which is run on ARM
(the existing DR_TRY test in client.file_io is not currently run on ARM).

Fixes #1845

Review-URL: https://codereview.appspot.com/287790043