i#1551 port to ARM: implement dr_clobber_retaddr_after_read() on ARM
Implements dr_clobber_retaddr_after_read() on ARM.  We change the
interface: the passed-in value is ignored and instead a value that is
guaranteed to not look like a return address is used (our stolen register
holding our TLS base, in fact).  This is for efficiency reasons, as on ARM
it would require an extra register spill in order to write an arbitrary
value.

Review-URL: https://codereview.appspot.com/284420043