i#1873 Android test suite: device copying and top-level suite
For copying binaries to an attached Android device, we use a strategy of
copying individual targets rather than a massive "adb push" at the very end
of the build.  While this does require individual copy commands scattered
through the cmake code, this runs much faster and uses much less space
(one-tenth) in addition to supporting "make test" and a much nicer
rebuild-and-rerun dev model,

Adds a new cmake option, DR_COPY_TO_DEVICE, which controls whether we copy
binaries.  Adds copy commands only for those targets needed for the current
tests in the limited Android test suite.

Renames DEVICE_BASEDIR to DR_DEVICE_BASEDIR to make it easier for
containing projects including DR to avoid name conflicts and to separate
options.

Adds Android builds to the top-level runsuite.  If "adb" is found and there
is an attached device, COPY_TO_DEVICE is turned on and test running is enabled.
For cron job suites, $ANDROID_TOOLCHAIN/bin should be added to the PATH.

Fixes #1873

Review-URL: https://codereview.appspot.com/287190043