i#1873 Android test suite: remote test execution
Adds test suite support on Android using the strategy of cross-compiling on
the host, copying the binaries over, and prefixing every command with "adb
shell".

A new cmake option, DEVICE_BASEDIR, defaulting to "/data/local/tmp",
controls where to copy.

Fixes minor build issues in the tests.

Not all ARM tests run yet: i#1874 covers fixing or disabling the failing
tests.

The automated copying of the binaries to the device is not yet in place.

Review-URL: https://codereview.appspot.com/290880043