i#1885: fix incorrect aflags entries for logical and multiply instrs
Our tables incorrectly had logical instructions as writing the
V flags and multiply instructions as writing V and C.

We also had logical instructions with immediates as blindly writing C when
in reality they leave C unchanged for small immediates.  We model this as
marking them as both reading and writing C.

Adds a small test.  Proper and thorough testing requires comparing to an
oracle implementation with flags information.

Fixes #1885

Review-URL: https://codereview.appspot.com/286550043