i#1906: ensure loads into PC use aligned addresses
Loading into the PC from an unaligned address is unpredictable and can lead
to crashes.  Our direct far links can be unaligned in Thumb mode.  We fix
this by inserting a 2-byte nop at the end of the fragment body to ensure
that all stubs, and thus all stub data addresses, are aligned.

Fixes #1906

Review-URL: https://codereview.appspot.com/294830043