i#1904: include genhtml in the release package
Adds an install rule for our modified version of the lcov genhtml script.
As it is under the GPL license, we also update the top-level license
documentation.

Augments the drcov docs to talk about our copy of genhtml, and to talk
about using -src_filter to avoid system library issues.

Fixes #1904

Review-URL: https://codereview.appspot.com/295780043