i#1569 AArch64: Implement parts of instr.c.
Implement functions to answer queries about instructions (is it a
call, does it read the thread register, and so on): the functions that
are called during execution of a simple program.

The MAX_FRAGMENT_SIZE on AArch64 is reduced to 0x8000 from the usual
default of 0xffff as this makes all branch instructions be in range.

Review-URL: https://codereview.appspot.com/297770043