i#1569 AArch64: Add pessimistic register operands when decoding OP_xx.
OP_xx is the opcode used for instructions that are not yet handled by
the decoder. With the exception of X30, written by BL and BLR, if an
A64 instruction accesses a general-purpose register then the number of
that register appears in one of four possible places in the
instruction word. So we can pessimistically assume that an unrecognised
instruction reads and writes all four of those registers, and this is
sufficient to enable correct (though often excessive) mangling.

Review-URL: https://codereview.appspot.com/294510043