i#1569 AArch64: Distinguish SP (stack pointer) from ZR (zero register).
Both are encoded as "register 31", so they are never alternatives in the
same encoding, but they can be alternatives in an alias. For example:

    mov x0, xzr    is an alias for    orr x0, xzr, xzr, lsl #0
    mov x0, sp     is an alias for    add x0, sp, #0, lsl #0

Therefore it is best to clearly distinguish them, and to prevent accidents,
make neither DR_REG_XZR nor DR_REG_XSP equal to (DR_REG_X30 + 1).

At the same time, remove the AArch64 register enum constants from AArch32.

Review-URL: https://codereview.appspot.com/298270043