i#1954: add better drreg support for intra-bb control flow
Adds on-by-default support for drutil_expand_rep_string() in drreg by
identifying the control flow and disabling lazy restores across app
instructions.

Adds drreg_set_bb_properties() with two flags, one for re-enabling
optimizations when app2app-inserted intra-bb control flow is compatible
with full drreg and one for disabling optimizations when
insertion-phase-inserted intra-bb control flow is incompatible with full
drreg lazy restores.

Adds documentation and a test, though the test is not perfect: it requires
manual tweaks for a full test of failure, which is hard to automate.

Fixes #1954

Review-URL: https://codereview.appspot.com/304730043