i#1990: avoid losing control in dr_app_running_under_dynamorio
When dr_app_running_under_dynamorio is invoked via a direct call (such as
when using DR as a static library), DR loses control.  This is because
dr_app_running_under_dynamorio is listed on must_not_be_inlined(), yet that
list means to leave as a native call.  The solution is to split the list
into leave_call_native() and must_not_be_entered().

A test will be included in the i#975 commit(s).

Fixes #1990

Review-URL: https://codereview.appspot.com/310730043