i#975 static DR: client existence and loading
Implements the proposed model for dynamorio_static: it allows any part of
the app to acts as the client, but also looks for and invokes a dr_init or
dr_client_main routine, supporting multiple methods of including a client
in the application.

Switches to accessing the client lib via the system loader when it's the
same as the application on Windows (the code for Linux was added in the
past).

Replaces checks for !IS_INTERNAL_STRING_OPTION_EMPTY(client_lib) with a new
macro CLIENTS_EXIST(), which is set to true for STATIC_LIBRARY.

Sets -code_api to true for STATIC_LIBRARY by default, and sets its
dependences via moving that code into options.c.

For STATIC_LIBRARY, "loads" the app itself as the client, and makes the
absence of an init routine non-fatal.

Updates the static_startstop test and adds two new tests: static_noclient
and static_noinit.

Review-URL: https://codereview.appspot.com/302630043