i#1989: crash calling dr_app_pc_from_cache_pc in thread exit event
We have to clean up the fragment state before we invoke the client's thread
exit event so that the client will see fragment deletion events before it
throws its own state away.  This means that API routines that need to
access fragment state (mainly those that need coudlbelinking status) will
not work in the thread exit event.  We document that and add checks for it.

Adds a test.

For CLIENT_INTERFACE builds, reduces the default max_bb_instrs from 1024 to
256 to avoid hitting the max size with heavyweight instrumentation clients.

Fixes #1989

Review-URL: https://codereview.appspot.com/303490043