i#95 Linux detach: translate do_syscall when method is sysenter
Today we do not translate the PC from do_syscall when the method is
sysenter.  This can happen if the thread just hasn't entered the kernel
yet, but due to i#2005 it is much more frequent in start/stop situations
where not all threads were updated to use sysenter.

This undoes a84214e, as the proper translation avoids the
is_dynamo_address() assert there.

Review-URL: https://codereview.appspot.com/310210043