i#1729 offline traces: add static bursty trace support
Adds testing for gathering a bursty trace using statically linked code:

Adds a static client drmemtrace_static.

Adds a test tool.drcacheoff.burst_static which includes an application that
uses the start/stop API to start DR midway through a big loop and to detach
fully from DR a few iterations later.  This "bursty trace" is saved by the
linked drmemtrace_static client to a file and passed to the cache simulator
as part of the test.

Review-URL: https://codereview.appspot.com/307330043