i#1727 drcachesim Windows: implement named pipes and fix build issues
Implements named pipes on Windows.  We still do not know the atomic pipe
write limit, and handling multiple processes connecting to the pipe is not
yet implemented as it is complex.  Single-process support is, however,
enough to enable building and testing drcachesim on Windows.

Fixes various build issues on Windows:
+ /MT vs /MTd problems due to sharing files between the client and exe
+ s/ERROR/ERRMSG/ as ERROR conflicts with Windows headers;
+ Misc warnings about casting, inttypes.h vs stdint.h, unreachable code,
  uninitialized variables, etc.

Places a copy of drinjectlib and drconfiglib into the drcachesim bin dir in
the build and install dirs.

Adds ERROR_PIPE_BUSY drwinapi support and uses it to check and issue a
NYI usage error for muliprocess online simulation on Windows.

Review-URL: https://codereview.appspot.com/305590044