i#1729 offline traces: implement initial optimized tracing
Implements the initial optimized offline tracing by defining a separate,
compact trace entry format that uses a single pointer-sized store for a
memref and a packed module entry for once-per-bb instruction entry.

This is being staged as several pieces and we disable the offline tests
here temporarily as the resulting file is not in the right format to be
read by the simulator: it needs postprocessing.

Still to come: module tracking and postprocessing.

Review-URL: https://codereview.appspot.com/309690043