i#1729 offline traces: add start+count to offline ifetch information
Eases the job of post-processing by adding the instruction count of each bb
into the offline trace entry.  This compresses the type down to 3 bits.

Switches to always using the first instruction in the bb for the ifetch
trace entry, to avoid having to decode backward during post-processing.

Adds special-case handling for repstr expansion to match the original code
stream's ifetch signature.

Review-URL: https://codereview.appspot.com/308630043