Support "drrun -t" with native tool offline actions
Removes the requirement that drrun take in an application, when a tool "-t"
is specified that has its own front-end.  This supports using the "-t"
syntax with post-processing or other offline actions.  This assumes that
other front-ends do their own checks for an application being present when
it needs to be.

Review-URL: https://codereview.appspot.com/306640043