i#95 Linux detach: do not eliminate threads or signal handlers
Moves dr_app_cleanup() a little closer to a detach: avoids terminating the
app's threads.  Documents that we expect the app to only call it near
process termination, and that dr_app_stop_and_cleanup() is the only full
detach we support right now.

Avoids clobbering the app's handlers at DR exit if we're detaching.

Review-URL: https://codereview.appspot.com/314820043