i#1729 offline traces: add timestamps and thread interleaving
Adds a new timestamp header for each buffer unit written to an offline
trace file.  The timestamp uses DR's new dr_get_microseconds() for a simple
cross-platform implementation.

Adds interleaving of threads in timestamp order to the raw2trace
post-processor.

Adds a test "burst-threads" which tests quite a few things:
+ Starting and stopping from a different thread than called dr_app_setup
  (i#2034, i#1921)
+ Detaching with many threads on Linux (i#2036)
+ Stress test for state translation with a heavyweight client (found
  i#2035)

The test is temporarily disabled on Windows until i#2040 is fixed.

The test produces such large trace files (in the other threads that keep
running while being synched for detach) that we cap their size with
-max_trace_size to keep this a small enough test for the regression suite.

Review-URL: https://codereview.appspot.com/317750043