i#1680 large pages: Add code to discover page size on Unix.
The stack pointer is passed as a third argument to relocate_dynamorio
and on Linux this is used to search the auxiliary vector for AT_PAGESZ.
Similarly, there are calls to os_page_size_init in dynamorio_app_init
and standalone_init to do the same thing. On MacOS, and on Linux if
for some reason os_page_size is called before these initialisations,
the page size is discovered using the system calls mmap and munmap,
which should work on any POSIX system.

A later commit should get Mac OSX to use sysctl_query on hw.pagesize.

Review-URL: https://codereview.appspot.com/313800043