i#1569 AArch64: Rename NUM_XMM_{REGS,SAVED,SLOTS} to NUM_SIMD_...
This is to prepare for moving some of x86/clean_call_opt.c into
clean_call_opt_shared.c.

Apart from the "#elif defined(X86)" in globals_shared.h, and adding
the definition of NUM_XMM_SLOTS for backward compatibility, this is
just search-and-replace with a few adjustments to spaces and comments.

Review-URL: https://codereview.appspot.com/318770043