i#2089 TLS init: add a safe read path for is_thread_tls_initialized()
This is a first step toward solving several issues with
is_thread_tls_initialized() by switching to using a new
safe_read_tls_base() step.  A new assembly routine and fault handling path
are added, along with globals to avoid the complexity of attempting this
during process init or exit.

This also fixes the TLS issues with the Windows Subsystem for Linux
(i#1986) and with other cases where it is very difficult to determine
whether TLS is initialized or not: it is much simpler to simply attempt to
read it.

The new code is under an off-by-default option -safe_read_tls_init until
we have the other pieces of the new TLS scheme in place.  Without the
forthcoming pieces, we incur a fault on every new thread init and exit.

Review-URL: https://codereview.appspot.com/313150043