i#2080 sigreturn: provide alternative to sigreturn for setcontext
Adds an on-by-default option -use_sigreturn_setcontext.  When turned off,
setting a thread's context will instead use dr_longjmp.  This is not fully
transparent: it assumes the target stack is valid and that we can clobber
the slot beyond TOS.  Using this instead of sigreturn is meant mainly as a
diagnostic to help debug future issues with sigreturn.

Review-URL: https://codereview.appspot.com/314150044