i#2006 generalize drcachesim: omit log msg when file ops replaced (#2177)
If the drmemtrace file ops are replaced, it's up the replacer to notify the
user where the trace data is going.  In some cases data is sent over the
network and the replaced create_dir is a nop that returns true, in which
case we don't want the existing message stating where the local log
directory is, as there is no such directory.