i#1621 AArch64 cc opt: Add test for -opt_cleancall=1. (#2395)
This patch adds a new cleancall-opt-1 test. It reuses most of the
infrastructure from inline.dll.c, which is moved to
cleancall-opt-shared.h and ported to AArch64.

cleancall-opt-1.dll.c at the moment only contains a single test function
(modify_gprs) which modifies all GPRs on AArch64, as well as the SIMD
registers on X86. Once #2333 is merged, an out-of-line clean call will
be used on AArch64. Currently no out-of-line call is created on X86,
as discussed in the pull request. This will be addressed in a separate
commit.