Handle ARM Linux kernel with CONFIG_IWMMXT enabled. (#2481)
With kernels up to 4.12, when CONFIG_IWMMXT is enabled the VFP frame
in the ucontext is offset by the size of the IWMMXT frame even when
no IWMMXT frame is written. We work around this by sending ourselves
a signal at init time and looking for the VFP frame in both places.