i#2681: AArch64 encoder: Consult check_reachable argument. (#2686)
A check_reachable member is added to the previously unused
decode_info_t struct and the encoder functions acquire an extra
argument which is a pointer to this struct. The functions that
encode branches, ADR and ADRP refer to this and accept unreachable
targets when check_reachable is false, setting the offset to zero
in this case. Targets are still required to be aligned.

Fixes #2681